Extract reducer to separate file
